Implemented 'count' keyword for parsing HGS datafiles
The 'count' keyword can greatly speed up reading HGS vector output data (like q, and v), as the size of the data to be read is know.

Refactored some of the pyhgs.parser.parse method's implementation to expose the expected type of data in any given type of HGS output file.