Enh: extract fluid mass bal. from transient sim .lst
Transient simulations have a slightly different text structure for the fluid mass balance block, including the potential for multiple mass balance blocks to appear in a timestep.

This enhancement will return the final fluid balance  block of the timestep.

Includes a test case and a lst file with 3 domains in the simulation, porous media, fracture, and well.