Fix: put copy of MC base_sim in CWD
The copy of the base simulation should be within the current working directory, not adjacent to base_sim. (base_sim could be very far away, directory-wise, so it make sense to have it nearer to the running process.)

Fix: include the base_sim's directory name as part of the temporary directory's name as intended.