Fix: solution times in output file
Solution times in the resultant zones were inconsistent with the solution times in the original data. This is because each newly read input dataset had its time assigned to the first zone in the output data. This was obviously incorrect, as the input data time should be applied to the current output zone.